app/requirements.txt
flask
flask-sqlalchemy
pymysql
werkzeug
pyjwt
pytest
pytest-flask
flask-login
cryptography
python-dotenv
PyJWT
gunicorn